Ajoute CollectivitePageLayout aux indicateurs
