+++ app.territoiresentransitions.react/src/app/pages/collectivite/Indicateurs/Indicateurs.tsx
@@ -1,10 +1,5 @@
-import {useParams} from 'react-router-dom';
-import {
-  indicateurIdParam,
-  indicateurViewParam,
-  IndicateurViewParamOption,
-} from 'app/paths';
-import {IndicateursNav, VIEW_TITLES} from './IndicateursNav';
+import {Link, useParams} from 'react-router-dom';
+import {indicateurIdParam} from 'app/paths';
 import {IndicateursPersonnalisesList} from './lists/IndicateursPersonnalisesList';
 import {IndicateursPredefinisList} from './lists/IndicateursPredefinisList';
 import {IndicateursClesList} from './lists/IndicateursClesList';
@@ -13,6 +8,44 @@ import {IndicateurPersonnalise} from './IndicateurPersonnalise';
 import {IndicateurPredefini} from './IndicateurPredefini';
 import {HeaderIndicateursList} from './Header';
 import IndicateurPersoNouveau from './IndicateurPersoNouveau';
+import {referentielToName} from 'app/labels';
+import {
+  indicateurViewParam,
+  IndicateurViewParamOption,
+  makeCollectiviteIndicateursUrl,
+} from 'app/paths';
+import CollectivitePageLayout from '../CollectivitePageLayout/CollectivitePageLayout';
+import {SideNavLinks} from 'ui/shared/SideNav';
+import {useCurrentCollectivite} from 'core-logic/hooks/useCurrentCollectivite';
+
+// correspondances entre item et libellé
+export const VIEW_TITLES: Record<IndicateurViewParamOption, string> = {
+  selection: 'Indicateurs prioritaires',
+  cles: 'Indicateurs clés',
+  //  tous: 'Tous les indicateurs',
+  perso: 'Indicateurs personnalisés',
+  cae: `Indicateurs ${referentielToName.cae}`,
+  eci: `Indicateurs ${referentielToName.eci}`,
+  crte: 'Indicateurs Contrat de relance et de transition écologique (CRTE)',
+};
+
+// items dans l'ordre de l'affichage voulu
+const ITEMS: IndicateurViewParamOption[] = [
+  'cles',
+  'selection',
+  'perso',
+  'cae',
+  'eci',
+  'crte',
+];
+
+// génère les liens à afficher dans la navigation latérale
+const generateIndicateursNavLinks = (collectiviteId: number): SideNavLinks => {
+  return ITEMS.map(indicateurView => ({
+    displayName: VIEW_TITLES[indicateurView],
+    link: makeCollectiviteIndicateursUrl({collectiviteId, indicateurView}),
+  }));
+};
 
 export const viewTitles: Record<IndicateurViewParamOption, string> = {
   ...VIEW_TITLES,
@@ -45,6 +78,8 @@ const IndicateurDetail = (props: {indicateurId: string; isPerso: boolean}) => {
  * détail d'un indicateur
  */
 const Indicateurs = () => {
+  const collectivite = useCurrentCollectivite();
+
   const params = useParams<{
     [indicateurViewParam]?: IndicateurViewParamOption;
     [indicateurIdParam]?: string;
@@ -53,15 +88,37 @@ const Indicateurs = () => {
   const view = params[indicateurViewParam] || 'perso';
   const isPerso = view === 'perso';
 
+  if (!collectivite) {
+    return null;
+  }
+
   return (
-    <div
-      className="fr-container !px-0 flex"
-      data-test={
-        indicateurId !== undefined ? `ind-${indicateurId}` : `ind-v-${view}`
-      }
+    <CollectivitePageLayout
+      sideNav={{
+        links: generateIndicateursNavLinks(collectivite.collectivite_id),
+        actions: !collectivite.readonly && (
+          <div className="mt-8">
+            <Link
+              data-test="create-perso"
+              className="fr-btn fr-btn--tertiary fr-ml-4w"
+              to={makeCollectiviteIndicateursUrl({
+                collectiviteId: collectivite.collectivite_id,
+                indicateurView: 'perso',
+                indicateurId: 'nouveau',
+              })}
+            >
+              Créer un indicateur
+            </Link>
+          </div>
+        ),
+      }}
     >
-      <IndicateursNav />
-      <div className="w-full">
+      <div
+        className="w-full"
+        data-test={
+          indicateurId !== undefined ? `ind-${indicateurId}` : `ind-v-${view}`
+        }
+      >
         {indicateurId !== undefined ? (
           <IndicateurDetail indicateurId={indicateurId} isPerso={isPerso} />
         ) : (
@@ -75,7 +132,7 @@ const Indicateurs = () => {
           </>
         )}
       </div>
-    </div>
+    </CollectivitePageLayout>
   );
 };
